Fixed typo and updated tests

+++ src/test/java/es/upm/etsit/dat/identi/RegistryTest.java
@@ -1,23 +1,24 @@
 package es.upm.etsit.dat.identi;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import java.util.List;
+
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 
 import org.apache.commons.lang3.RandomStringUtils;
-
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-
-import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
 import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -33,15 +34,20 @@ import org.springframework.web.context.WebApplicationContext;
 
 import es.upm.etsit.dat.identi.controllers.RegistryController;
 import es.upm.etsit.dat.identi.forms.CensusMemberForm;
+import es.upm.etsit.dat.identi.persistence.model.CDMember;
 import es.upm.etsit.dat.identi.persistence.model.CDToken;
 import es.upm.etsit.dat.identi.persistence.model.CensusMember;
+import es.upm.etsit.dat.identi.persistence.model.Commission;
 import es.upm.etsit.dat.identi.persistence.model.CommissionToken;
+import es.upm.etsit.dat.identi.persistence.model.CommissionMember;
+import es.upm.etsit.dat.identi.persistence.repository.CDMemberRepository;
 import es.upm.etsit.dat.identi.persistence.repository.CDTokenRepository;
 import es.upm.etsit.dat.identi.persistence.repository.CensusMemberRepository;
 import es.upm.etsit.dat.identi.persistence.repository.CommissionRepository;
 import es.upm.etsit.dat.identi.persistence.repository.CommissionTokenRepository;
 import es.upm.etsit.dat.identi.persistence.repository.DegreeRepository;
-import es.upm.etsit.dat.identi.persistence.repository.DepartamentRepository;
+import es.upm.etsit.dat.identi.persistence.repository.DepartmentRepository;
+import es.upm.etsit.dat.identi.persistence.repository.CommissionMemberRepository;
 
 @SpringBootTest(classes = IdentidatApplication.class)
 @AutoConfigureMockMvc
@@ -59,7 +65,7 @@ public class RegistryTest {
     private CDTokenRepository cdTknRepo;
 
     @Autowired
-    private DepartamentRepository dptRepo;
+    private DepartmentRepository dptRepo;
 
     @Autowired
     private CommissionTokenRepository cmmTknRepo;
@@ -73,14 +79,17 @@ public class RegistryTest {
     @Autowired
     private DegreeRepository dgrRepo;
 
+    @Autowired
+    private CDMemberRepository cdMemRepo;
+
+    @Autowired
+    private CommissionMemberRepository cmmMemRepo;
+
     @Mock
     DefaultOAuth2User principal;
 
-    private static String cdTokenString = RandomStringUtils.randomAlphanumeric(20);
-    private static CDToken cdToken;
-
-    private static String cmmTokenString = RandomStringUtils.randomAlphanumeric(20);
-    private static CommissionToken cmmToken;
+    private static String cdTokenString;
+    private static String cmmTokenString;
 
     private String givenName = "Perico";
     private String familyName = "Pérez Pérez";
@@ -89,6 +98,12 @@ public class RegistryTest {
     private String personalID = "73745001D";
     private String phone = "666666666";
     private String degreeCode = "09DA";
+
+    @BeforeAll
+    public static void setTokenStrings() {
+        cdTokenString = RandomStringUtils.randomAlphanumeric(20);
+        cmmTokenString = RandomStringUtils.randomAlphanumeric(20);
+    }
 
     @BeforeEach
     public void generatePrincipal() {
@@ -107,15 +122,17 @@ public class RegistryTest {
         when(principal.getAttribute("family_name")).thenReturn("Pérez Pérez");
     }
 
-    @Test
+    @BeforeEach
     public void contextLoads() throws Exception {
         assertThat(registryController).isNotNull();
     }
 
     @Test
     public void checkRegisterForm() throws Exception {
-        cdToken = new CDToken(cdTokenString, dptRepo.findByAcronym("DIT"));
+        CDToken cdToken = new CDToken(cdTokenString, dptRepo.findByAcronym("DIT"));
         cdTknRepo.saveAndFlush(cdToken);
+
+        assertNotNull(cdTknRepo.findByToken(cdTokenString));
 
         this.mockMvc.perform(get("/register?token=" + cdTokenString)).andExpect(status().isOk())
                 .andExpect(content().string(org.hamcrest.Matchers.containsString("Registro en el censo de DAT")));
@@ -123,6 +140,11 @@ public class RegistryTest {
 
     @Test
     public void checkFormSubmission() throws Exception {
+        CDToken cdToken = new CDToken(cdTokenString, dptRepo.findByAcronym("DIT"));
+        cdTknRepo.saveAndFlush(cdToken);
+
+        assertNotNull(cdTknRepo.findByToken(cdTokenString));
+
         CensusMemberForm cenMemForm = new CensusMemberForm(this.givenName, this.familyName, this.email, this.degreeCode,
                 TokenType.CD, cdTokenString);
         cenMemForm.setPersonalID(this.personalID);
@@ -132,25 +154,61 @@ public class RegistryTest {
         this.mockMvc.perform(post("/register").sessionAttr("censusMemberForm", cenMemForm))
                 .andExpect(status().isFound());
 
-        cdTknRepo.delete(cdToken);
-        cenMemRepo.delete(cenMemRepo.findByUsername(this.username));
+        CensusMember censusMember = cenMemRepo.findByUsername(this.username);
+        assertEquals(this.givenName, censusMember.getName());
+        assertEquals(this.familyName, censusMember.getSurname());
+        assertEquals(this.personalID, censusMember.getPersonalID());
+        assertEquals(this.username, censusMember.getUsername());
+        assertEquals(this.email, censusMember.getEmail());
+        assertEquals(this.phone, censusMember.getPhone());
+        assertEquals(dgrRepo.findByCode("09DA"), censusMember.getDegree());
+
+        List<CDMember> cdPositions = cdMemRepo.findByCensusMember(censusMember);
+        assertEquals(1, cdPositions.size());
+        assertEquals(dptRepo.findByAcronym("DIT"), cdPositions.get(0).getDepartment());
     }
 
     @Test
     public void checkPositionAddition() throws Exception {
-        cmmToken = new CommissionToken(cmmTokenString, cmmRepo.findByName("Junta de Escuela"));
+        Commission je = cmmRepo.findById(Long.valueOf(1)).get();
+        CommissionToken cmmToken = new CommissionToken(cmmTokenString, je);
         cmmTknRepo.saveAndFlush(cmmToken);
 
+        assertNotNull(cmmTknRepo.findByToken(cmmTokenString));
+
         cenMemRepo.saveAndFlush(new CensusMember(Long.valueOf(66), this.givenName, this.familyName, this.email,
-                this.username, this.personalID, this.phone, dgrRepo.findByCode(this.degreeCode), null, null));
+        this.username, this.personalID, this.phone, dgrRepo.findByCode(this.degreeCode), null, null));
 
         CensusMemberForm cenMemForm = new CensusMemberForm(this.givenName, this.familyName, this.email, this.degreeCode,
                 TokenType.COMMISSION, cmmTokenString);
-                
+
         this.mockMvc.perform(post("/register").sessionAttr("censusMemberForm", cenMemForm)).andDo(print())
                 .andExpect(status().isFound());
 
-        cmmTknRepo.delete(cmmToken);
-        cenMemRepo.delete(cenMemRepo.findByUsername(this.username));
+        List<CommissionMember> cmmPositions = cmmMemRepo.findByCensusMember(cenMemRepo.findByUsername(this.username));
+        assertEquals(1, cmmPositions.size());
+        assertEquals(je, cmmPositions.get(0).getCommission());
     }
+
+    @AfterEach
+    public void cleanDB() {
+        CensusMember censusMember = cenMemRepo.findByUsername(this.username);
+        if (censusMember != null) {
+            cenMemRepo.delete(censusMember);
+            cenMemRepo.flush();
+        }
+
+        CDToken cdToken = cdTknRepo.findByToken(cdTokenString);
+        if (cdToken != null) {
+            cdTknRepo.delete(cdToken);
+            cdTknRepo.flush();
+        }
+
+        CommissionToken cmmToken = cmmTknRepo.findByToken(cmmTokenString);
+        if (cmmToken != null) {
+            cmmTknRepo.delete(cmmToken);
+            cmmTknRepo.flush();
+        }
+    }
+
 }
